+++ app/src/main/java/com/kiero/presentation/main/navigation/MainAppState.kt
@@ -34,8 +34,6 @@ import com.kiero.presentation.parent.navigation.ParentReward
 import com.kiero.presentation.parent.navigation.ParentSchedule
 import com.kiero.presentation.parent.screen.alarm.navigation.navigateToAlarm
 import com.kiero.presentation.parent.screen.journey.navigation.navigateToParentJourney
-import com.kiero.presentation.parent.screen.mission.navigation.navigateToMissionAdd
-import com.kiero.presentation.parent.screen.mission.navigation.navigateToParentMission
 import com.kiero.presentation.parent.screen.mypage.navigation.navigateToMypage
 import com.kiero.presentation.parent.screen.mission.directadd.navigation.navigateToMissionAdd
 import com.kiero.presentation.parent.screen.mission.overview.navigation.navigateToParentMission
