+++ app/src/main/java/com/kiero/presentation/parent/navigation/ParentNavigation.kt
@@ -10,13 +10,13 @@ import androidx.navigation.compose.navigation
 import com.kiero.core.navigation.Route
 import com.kiero.presentation.parent.screen.alarm.navigation.parentAlarmNavGraph
 import com.kiero.presentation.parent.screen.journey.navigation.parentJourneyNavGraph
-import com.kiero.presentation.parent.screen.mission.ParentAddMissionRoute
-import com.kiero.presentation.parent.screen.mission.auto.navigation.navigateToAutoMissionAdd
-import com.kiero.presentation.parent.screen.mission.auto.navigation.parentAutoMissionAddNavGraph
-import com.kiero.presentation.parent.screen.mission.navigation.MissionEdit
-import com.kiero.presentation.parent.screen.mission.navigation.navigateToMissionAdd
-import com.kiero.presentation.parent.screen.mission.navigation.parentMissionAddNavGraph
-import com.kiero.presentation.parent.screen.mission.navigation.parentMissionNavGraph
+import com.kiero.presentation.parent.screen.mission.autoadd.navigation.navigateToAutoMissionAdd
+import com.kiero.presentation.parent.screen.mission.autoadd.navigation.parentAutoMissionAddNavGraph
+import com.kiero.presentation.parent.screen.mission.directadd.ParentAddMissionRoute
+import com.kiero.presentation.parent.screen.mission.directadd.navigation.MissionEdit
+import com.kiero.presentation.parent.screen.mission.directadd.navigation.navigateToMissionAdd
+import com.kiero.presentation.parent.screen.mission.directadd.navigation.parentMissionAddNavGraph
+import com.kiero.presentation.parent.screen.mission.overview.navigation.parentMissionNavGraph
 import com.kiero.presentation.parent.screen.mypage.navigation.parentMypageNavGraph
 import com.kiero.presentation.parent.screen.reward.navigation.parentRewardNavGraph
 import com.kiero.presentation.parent.screen.schedule.navigation.parentScheduleNavGraph
